+++ packages/components/Menu/src/MenuTrigger/MenuTrigger.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
-import { memoize, stagedComponent } from '@fluentui-react-native/framework';
-import { menuTriggerName, MenuTriggerChildProps, MenuTriggerState } from './MenuTrigger.types';
+import { stagedComponent } from '@fluentui-react-native/framework';
+import { menuTriggerName } from './MenuTrigger.types';
 import { useMenuTrigger } from './useMenuTrigger';
-import { AccessibilityActionEvent } from 'react-native';
 import { MenuTriggerProvider } from '../context/menuTriggerContext';
-import { InteractionEvent, isMouseEvent } from '@fluentui-react-native/interactive-hooks';
+import { getRevisedProps } from './getRevisedProps';
 
 export const MenuTrigger = stagedComponent((_props: React.PropsWithChildren<Record<never, any>>) => {
   const menuTrigger = useMenuTrigger();
@@ -29,49 +28,5 @@ export const MenuTrigger = stagedComponent((_props: React.PropsWithChildren<Reco
   };
 });
 MenuTrigger.displayName = menuTriggerName;
-
-const getRevisedProps = memoize(getRevisedPropsWorker);
-function getRevisedPropsWorker(state: MenuTriggerState, props: any): MenuTriggerChildProps {
-  const revisedProps = { ...state.props };
-  if (props.accessibilityState) {
-    revisedProps.accessibilityState = { ...state.props.accessibilityState, ...props.accessibilityState };
-  }
-
-  if (props.accessibilityActions) {
-    revisedProps.accessibilityActions = [...state.props.accessibilityActions, ...props.accessibilityActions];
-  }
-
-  if (props.onAccessibilityAction) {
-    revisedProps.onAccessibilityAction = (e: AccessibilityActionEvent) => {
-      props.onAccessibilityAction(e);
-      state.props.onAccessibilityAction(e);
-    };
-  }
-
-  if (props.onClick) {
-    revisedProps.onClick = (e: InteractionEvent) => {
-      props.onClick(e);
-      state.props.onClick(e);
-    };
-  }
-
-  let onHoverIn = undefined;
-  if (props.onHoverIn) {
-    onHoverIn = (e: InteractionEvent) => {
-      props.onHoverIn(e);
-      state.props.onHoverIn(isMouseEvent(e) && e);
-    };
-  }
-
-  let onHoverOut = undefined;
-  if (props.onHoverOut) {
-    onHoverOut = (e: InteractionEvent) => {
-      props.onHoverOut(e);
-      state.props.onHoverOut(isMouseEvent(e) && e);
-    };
-  }
-
-  return { ...revisedProps, onHoverIn, onHoverOut };
-}
 
 export default MenuTrigger;
